Introduces rudimentary ping-pong protocol
The purpose of this commit is to implement a basic websocket behavoiur.
We will expect a handshake request from the client, process it according
to standard procedure and enter a process which is observable as some
type of record of what happened. This is just a unit for timestamp at
the moment.