Read a frame of data into bytesequences
We use a statemachine that switches state depending on bytes it
encounters. The crucial bit is to parse the length of the payload to
tell where one frame starts and another ends. To get the tally right we
also need to account for absence/precence of the four mask bytes and the
varying length of the bytes that encodes the payload length.