Transforms frames to messages.
This commit also has quite a lot of renames et.c. What we are trying to
achieve is a clearer logic as to what reaction belongs to what frame
that is sent over the wire What we are trying to achieve is a clearer
logic as to what reaction belongs to what frame that is sent over the
wire.