Rewrite writer
The idea is to project two things from the connectedSocket: a reader and
a writer. The reader exposes an observable of read buffers from the
stream tagges with their socket identifier. The writer exposes a
function that takes an observable of buffers and return a timestamped
socket identifier. This represents that a buffer was written to the
socket at that time.