Makes invalidframe messages
We need to rewrite the pipeline from byte to message. The way it is is
really messy. We also want to have a place int the pipeline after we
generate the messages to introduce some state in the messageflow such as
ping waiting for pong and close handshake. We shall see if we can make
testcase 2.5 pass without this.