Adds types and function signatures for defrag
We do framing in three steps, parsing, validation and defragmentation.
This commit deals with the defragmenting. The defragmentation
will receive parsed or bad frames and transform these to bad or
defragmented frames. That is, leave the bad frames as they were then
evaluate each parsed frame. if it has a continuation we need to buffer
until the final frame arrives. While waiting for continuation we must
not receive another frame that is the beginning of a continuation. Also
we must make sure that the client only sends text and binary messages
fragmented. At that time it will emit a defragmented
frame. Else it will emit a badframe. Frames that are final will
be emitted as defragmented frames.