Parses utf8 stream oriented
We can handle that utf8 is split into different frames so that two
frames contains different parts of the same codepoint.

We found a bug in the parsing when payload is 10 bytes long. We will
siplify the logic to assume masked frames at all times. According to the
spec a server should end communication immediately with a client that
starats sending unmasked frames anyway.