Frame-parse bails out on unmasked frame
According to the spec a server should disconnect a client that sends an
unmasked frame. Therefore there's no reason to hav logic server-side to
parse unmasked frames. This commit removes this logic from the parsing.
Instead we always assume a frame has the msk bytes. If the bitfield says
that the frame is unmasked we report this to the function that
constructs our frame-objects so that it can emit a Badframe that will
trigger a shutdown. Something to consider is to do all the bitfield
validation already in the parsing and bail-out there.