utf8-state handles two-char unicode chars
.net represents codepoints that continues over more than two bytes as
two consecutive chars. The utf8-state assumed one char per unicode.

Introduce Pingpong

We want to respond incoming ping-messages with pong. Pong messages
should be passed to a function that returns a pingmessage or a
closemessage. These should be passed to the output.

Introduces OpcodeMessage

The plan is to have like a straight pipeline from reading socket to
writing back to it. We map frames to eiter textmessage, binarymessage or
opcodemessage. We map these to a steam of bytearrays for the socket to
write. The messaging should accept one function to map textmessages and
one to map binarymessages. We should also accept a function that take a
stream of pongmessages and returns a stream of pingmessages.

Handshakes

- write test for parson to first parse one handshakerequest and then
continue parsing frames repeatedly

Output

Todo:
- Rename Framing, messaging and socketing to Protocol, Application and
IO. Prefer technical names.
- Make sure the main application works with the fuzzer.
- Make sure all functions that needs scheduling can be passed a
scheduler. Deserialization is done.
- Rewrite or delete ignored tests

Moves tomessage to protocol

The idea is to move all functions out of the application layer. The only
thing the applicationlayer will do is receive a stream of data-frames,
connectionclose, handshakeaccepts or client pongs and is expected to produce a stream of output (dataframes, handshakeresponse or pings)

Adds Logging to file

Closehandshake

removed schedulers from Handshake

Fuzzer working