Pingpong
The idea is to use incoming pongs to create windows and return both the
latency metric (For Logging etc.) and Pings (IO-bound). We accept a
timespan from the program to generate a stream of server-side pings.

Need to look into if the ping/pong really starts until we get a message.
The behaviour should be that the ping pong starts after the handshake.