Introduce parse type
Parse is meant to represent an incoming message as it goes throuch the
protocol pipeline. It is either a successful parse or an error. The
purpose is to make the types that represents a successful parse more
coherent. Another motivation is to be able to handle errors in the
parsing together with incoming closeframes in some closehandshake-logic.
The alternative to this solution would have been to introduce separate
our close type. The poor pattern-matching features of c# makes this
solution not so nice as it could have been. Especially when used
together with observable create or other operators that deals directly
with the obesrvable monad.