Removes TextParse type
I decided to let it be up to the application to represent the payload as
strings. The servers responsibility is to validate that the payload is
valid utf8 when the opcode is text.

The application must be expressed in a transformation of a stream of
messages which are a combination of opcode (always text or binary) and payload (which in case opcode is 1 is guaranteed to be valid utf8)

The server should make it possible to send continuations from the
application.