Adds rudimentary cli args parsing
We look for arguments in the first and second position. An argument that
looks vaguely like an ipv4-adress in the first (does not check for
invalid range et.c. just for basic for dot separated positive integers below 999) and an argument that can be parsed to an int in the second for the port number. Parameters default to 0.0.0.0:80 which is how we want to launch in the docker container.