Excludes unit-test project from Dockerfile
In it s current state wsr behaves differently when we connect to it vie
loopback interface than when we onnect to it running in a container. As
of now I don¨t know the reason for this. When run in a container Wsr
performs poorly timing out on long frames. Also there is strange errors
that cannot be reproduced reliably (i.e. test sometimes pass, sometimes
not). Examples of these are

- missing first frame after handshake
- corrupt incoming bytes.

Things I've tried

- remove schedulers from the FromAsync calls that converts
socket-methods to observables.
- downgrade to netcoreapp1.1.1/sdk1.1.4