Addresses performance issue with FrameBytestate
Problem is that parsing a long stream of bytes is to slow. What took
almost no noticeable time in the old implementation now takes about 3s.
What I did that had an effect on performance:
- Implemented IEquateable on the data-types.
- Made the test not call to string on every framebyte unless it fails.
- Took away one ulong prop from the framebyte struct
What I did that did not have significant effect:
- changed Either from struct to class.
Todo:
- rewrite the readpayload function so that decrements a variable in a
closure to index the mask-array instead of recursion.