Types should not be structs
I had this idea that it would make sense to define the types as structs
because they are so short-lived and small. Im going to stick with
classes since that is the default choice and the types ended up being
fairly complex. Changing to classes also make the test run a bit faster
which indicates that there is a noticeable direct performance penalty to
using structs here. when I have better ways to measure the code I might
go back and experiment.