Adds operator to echo incoming frames
The minimal implementation of the echoserver is to read incoming bytes
and starting with a handshake produce a mapping from incoming bytestream to
outgoing bytebuffers.

The handshake needs to determine request length an validity then produce
a valid accept or decline-message.

The frame needs to calculate
1. frame length
2. opcode
3. unmasked payload bytes
4. recombine to a conformant server side ws-frame

ToDo find a way to have frame follow on handshake like it should.