Handshake rewrite
We parse the handshake in two steps:
1. Buffer the bytes by the clrf-sequences. An upgrade-request has no
payload so we it will always end in one empty buffer so we condition the
completion of this sequence on this.
2. Use these buffers to build up a representation of the request that is
suitable for logging and calculating an appropriate response.
3. the logic for the response is in this commit as basic as it can be.
We assume the request is always well-formed and possible to satisfy. It
uses the request key and calculates a response-key according to the
specification. We serialize this as a http-message.