Passes first fuzzertest
Wsr takes as an input a routingtable that maps a url to a way of
transforming a stream of bytes to a stream of outputbuffers.

Wsr offers functions to construct a route by supplying a way to produce
a stream of opcode bytestream pairs by subscibing to the incoming opcode
bytestream pairs.