Passes fuzzer 2.* cases
We route the appdata-stream first by frameId, then by Opcode.

We detect invalid conditions in the bytestream by testing and throwing
exceptions. These exeptions emits as OnError messages when we subscribe to
the appdata-stream. This means at the latest in the
Serialization-projection. Here we catch the exceptions and map them to a
proper app-data for a close frame and end the stream.

Use default behavior of not sending own pings