Removes guid
I introduced a guid for each frame to have something to group by. Now I
realize that I don't need this because of how group by until is working.
To group by opcode and complete the stream on a change in a flag is
enough for what we need to do. This is good because now we can describe
a frambyte i three bytes, the application data, the opcode and a byte
for socket internal use. this is three bytes.