Handles continuations
We will handle continuations by relying on the workings of
groupbyuntil. A frame in the context of the appdatastream shares the
same opcode and emits a oncompleted when i reaches a byte that have
final bit set in its controlbyte. We also store there Infomation on the opcode of a sequence of continuationframes.

Simple continuation

We use the control-byte to keep track of continuations.