Passes testcase 7.*
It is important at what time we subscribe to the appdata stream. If we
do it after the complete on close behaviour the stream won't complete in
order since then its the timing of the first byte that determines the
order of the frames and not the last appdata byte. Hence an interleaved
closeframe comes after the frame it is interleaved in which is not
according to spec.