Try to increase performance of deserialize
Calling it failure at this point. I need to know more about what causes
the performance-problem. Is it with IO or is it with processing. My
suspicion is that it is with IO. I'm benchmarking against the kestel
server that is known for its optimized io and wsr uses a .net
TCP-socket class that wraps their socket class. It also mentions that
high performance io might want to deal with the socket directly. I tried
an expriment where I copied some of the code in the rxx project into
this one to see if that code could get me faster io. I tested it out and
I think it did. However it corrupted all frames that did not fit into
the buffer size and other wierd problems. I'll leave this experiment in
the code but consider this branch finished.