Rewrite deserialize to improve performance.
This version of WSr is about three times as fast as the old one. Still
its between 2 to 10 times slower than kestrel websocket. This commit
includes a perf test. From this we can se that the server logic handles
a 64M size payload in about 1000 ms. This is three times as slow as when
receiving this payload size through the socket. IO adds about 2000 ms to
the time here.